pyproject.toml
[tool.poetry]
name = "e2e-testing"
version = "0.1.0"
description = "Pantos End to End tests"
authors = ["Pantos GmbH <contact@pantos.io>"]
license = "GPL-3.0-only"
readme = "README.md"
package-mode = false

[tool.poetry.group.test.dependencies]
pytest = "8.2.2"
pytest-cov = "5.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.lint.dependencies]
bandit = "1.7.9"
flake8 = "7.1.0"
isort = "5.13.2"
mypy = "1.10.1"
types-pyyaml = "6.0.12.20240311"
types-requests = "2.32.0.20240712"
yapf = "0.40.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.7.1"

[tool.poetry.dependencies]
python = "^3.12"
pantos-client-library = "^1.1.4"
pytest-ordering = "^0.6"
pytest-xdist = "^3.6.1"
pytest-timeout = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
